Change models to use normalized short texcoords

+++ include/Engine/Gfx/ModelReader.hpp
@@ -16,7 +16,7 @@ namespace Engine::Gfx {
 		Vertex() {};
 
 		glm::vec3 pos;
-		glm::vec2 uv;
+		glm::u16vec2 uv;
 
 		// TODO: instead of invalid bone id we can just check if weight == 0, can do this in shader also
 		uint8 bones[4];
